+++ examples/server_hardware/data_source.tf
@@ -1,0 +1,16 @@
+provider "oneview" {
+  ov_username =   "${var.username}"
+  ov_password =   "${var.password}"
+  ov_endpoint =   "${var.endpoint}"
+  ov_sslverify =  "${var.ssl_enabled}"
+  ov_apiversion = 2600
+  ov_ifmatch = "*"
+}
+
+data "oneview_server_hardware" "server_hardware" {
+	name = "0000A66102, bay 4"
+}
+
+output "oneview_server_hardware_value" {
+	value = "${data.oneview_server_hardware.server_hardware.uri}"
+}