+++ CHANGELOG.md
@@ -1,6 +1,47 @@
 All notable changes to this project will be documented in this file . 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+# [v6.0.0-11] (Unreleased)
+### Notes
+- This release supports API2600 minimally where we can use OneView v6.00 with this SDK.
+
+#### Major changes
+
+### Oneview Features supported
+- Ethernet Network
+- Enclosure
+- Enclosure Group
+- FC Network
+- FCOE Network
+- Hypervisor Manager
+- Hypervisor Cluster Profile
+- Interconnects
+- Interconnect Types
+- Logical Enclosure
+- Logical Interconnects
+- Logical Interconnect Groups
+- Network Set
+- Scope
+- Server Certificate
+- Server Hardware
+- Server Hardware Type
+- Server Profile
+- Server Profile Template
+- Storage Pool
+- Storage Volume
+- Storage Volume Attachments
+- Storage Volume Template
+- Storage System
+- Uplink Set
+
+### Bug fixes & Enhancements:
+
+- [#238] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/238) server profile connection settings
+- [#233] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/233) terraform import oneview_server_profile_template missing settings.
+- [#234] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/234) local storage configuration.
+- [#242] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/242) serverprofile bootmode_secureboot.
+- [#267] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/267) local_storage not working as expected.
+- [#268] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/268) local storage issue ( defined in server profile template). 
 
 # [v1.7.0]
 ### Notes
@@ -36,14 +77,6 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Storage Volume Template
 - Storage System
 - Uplink Set
-
-### Bug fixes & Enhancements:
-- [#233] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/233) terraform import oneview_server_profile_template missing settings.
-- [#238] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/238) server profile connection settings
-- [#234] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/234) local storage configuration.
-- [#242] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/242) serverprofile bootmode_secureboot.
-- [#267] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/267) local_storage not working as expected.
-- [#268] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/268) local storage issue ( defined in server profile template). 
 
 # [v1.6.0]
 ### Notes